refactor(#6): User관련 Exception 패키지 경로 수정 - exception 패키지의 위치를 application.user.port.in.exception에서 application.user.exception으로 수정했어요.
