+++ src/main/java/com/teka/application/user/exception/EmailAlreadyExistsException.java
@@ -1,6 +1,6 @@
-package com.teka.application.user.port.in.exception;
+package com.teka.application.user.exception;
 
-import com.teka.application.user.port.in.exception.error.UserApplicationErrorProperty;
+import com.teka.application.user.exception.error.UserApplicationErrorProperty;
 import com.teka.shared.error.TekaException;
 
 public class EmailAlreadyExistsException extends TekaException {
